Move proxy.frontend.command package (#7297)
* Update javadoc

* Update javadoc

* Add generic of MySQLProtocolFrontendEngine

* Add generic of DatabasePacketCodecEngine

* Move proxy.frontend.command package
