+++ shardingsphere-proxy/shardingsphere-proxy-frontend/shardingsphere-proxy-frontend-spi/src/main/java/org/apache/shardingsphere/proxy/frontend/command/CommandExecutor.java
@@ -15,51 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.frontend.api;
+package org.apache.shardingsphere.proxy.frontend.command;
 
 import org.apache.shardingsphere.db.protocol.packet.DatabasePacket;
 
 import java.sql.SQLException;
+import java.util.Collection;
 
 /**
- * Query command executor.
+ * Command executor.
  */
-public interface QueryCommandExecutor extends CommandExecutor {
+public interface CommandExecutor {
     
     /**
-     * Judge is query SQL or not.
+     * Execute command.
      *
-     * @return is query SQL or not
-     */
-    boolean isQueryResponse();
-    
-    /**
-     * Judge is update response.
-     *
-     * @return is update response or not
-     */
-    boolean isUpdateResponse();
-    
-    /**
-     * Judge is error response.
-     *
-     * @return is error response or not
-     */
-    boolean isErrorResponse();
-    
-    /**
-     * Goto next result value.
-     *
-     * @return has more result value or not
+     * @return database packets to be sent
      * @throws SQLException SQL exception
      */
-    boolean next() throws SQLException;
-    
-    /**
-     * Get query data.
-     *
-     * @return database packet of query data
-     * @throws SQLException SQL exception
-     */
-    DatabasePacket<?> getQueryData() throws SQLException;
+    Collection<DatabasePacket<?>> execute() throws SQLException;
 }
